--- src/scenes/ViewInquests/components/ViewInquests/index.js
@@ -1,1 +1,0 @@
-export { default } from './ViewInquests.js';
